Add a portable shell install script
This is not quite POSIX shell, we do ask for `local` to be available,
and a few basic commands.  It's about as portable as I can make it
without twisting into deep knots which risk adding bugs.

External commands required beyond POSIX: curl unzip mktemp

Document the usage of this script.

Adjust the Python to also use ~/.bashrc not ~/.bash_profile because the
profile won't be run in all situations while the rc file should be used
whenever the user has an interactive bash shell.

We add symlinks into ~/bin if it exists, and if that is in `$PATH`
already then we avoid the "how to add our install dir to PATH"
instructions; implemented for the Python too for consistency.

We hard-code a pattern of download URLs into the script, instead of
hard-coding the pattern of API server URLs.  This might even be _more_
stable longer-term, but if GitHub change the URL schema then we can
change the install script.  Also embed the same patterns into the
Python script instead of using regular expressions to parse
GitHub-generated HTML.  (Earlier versions of this PR used jq(1) and the
API system, jq dropped for portability concerns.)

Usage for the shell differs slightly from the Python; it has a `-h` help
option:

    Usage: install [-t <tag>] [-d <dir>] [-s <dir>]
     -d dir     directory to download into [default: ~/.nsccli/bin]
     -s dir     directory in which to place a symlink to the binary
                [default: ~/bin] [use '-' to forcibly not place a symlink]
     -t tag     retrieve a tagged release instead of the latest