Added env variables to prevent creation of gitignore, performing self-update checks, or writing ~/.nsc.json.
NSC_CWD_ONLY - when set the context is based on location of the cwd (operator/account).
NSC_NO_SELF_UPDATE - when set will skip any self-update checks
NSC_NO_GIT_IGNORE - when set will not write .gitignore files in the directory pointed to by $NKEYS_PATH. If the nkeys are in a directory under revision control, it means that the secrets will be checked in. Exercise caution.

FIX #240