[added] operator option require-signing-keys and simplify signing keys
The option require-signing-keys requires subordinate objects to be
signed by signing keys only.
In this mode identity keys are not suggested for signing.

Where the option --sk (for signing key) can be used, the value
"generate" causes a signing key to be generated/stored on the fly"

Updated jwt library

Signed-off-by: Matthias Hanel <mh@synadia.com>