[FIX] removed misleading flag help for `service` as only supported on public imports [FEAT] Added (hidden) support for `--account-token-position` and `--advertise` for public exports.
The `--account-token-position` allows public exports from an account to be constrained to
a subjects that are clamped to the account. The NATS server will ensure that only an
importing account sporting a matching public key in the specified wildcard token can
access the service/stream. This removes the burden of generating a token for subjects
that follow such a pattern on system-wide services