sync goreleaser targets to include most natscli ones
I use FreeBSD.  We build the natscli repo for FreeBSD, let's build nsc for
FreeBSD too.  In fact, let's pull in the build combinations from the
`.goreleaser.yml` in the natscli repo.

We also build for Windows and ignore arm64 there, where natscli doesn't; I
don't know enough to be confident here, so opted to not add new Windows builds
with this PR.

Resulting binary tested on FreeBSD.  Works nicely.