releng: remove: ssh; signing keys from disappearing disk
Wrap both signing commands in consistent wrapper shims.  These will expose the
signing keys in the right ways to the commands.

Trigger release flow for 0.0.x too, for testing.

Temporarily stop using the ssh signer, because of bugs in the signing flow with
older OpenSSH, fixed in more recent releases; we'll try again with 22.04.