Lock goreleaser version, adjust name templates; license
In debugging missing architectures from the generated brew data,
it eventually was "goreleaser bug, fixed in newer patchlevel release".

We had one bug in that the archive names for 32-bit ARM would have a vN suffix
but the specification for brew did not include that.  Fixed, and include a
reference to the default and just leave all the architecture finangling
in-place in both.

nsc is under the Apache-2.0 license (SPDX identifier), so tell Brew that.